+++ nucliadb/nucliadb/export_import/__init__.py
@@ -17,3 +17,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
+import logging
+
+SERVICE_NAME = "nucliadb.export_import"
+logger = logging.getLogger(SERVICE_NAME)
