+++ nucliadb/nucliadb/export_import/utils.py
@@ -17,12 +17,15 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
-from typing import AsyncGenerator, Callable
+from io import BytesIO
+from typing import AsyncGenerator, AsyncIterator, Callable, Optional
 
 from nucliadb.common.context import ApplicationContext
 from nucliadb.common.datamanagers.entities import EntitiesDataManager
 from nucliadb.common.datamanagers.labels import LabelsDataManager
 from nucliadb.common.datamanagers.resources import ResourcesDataManager
+from nucliadb.export_import.exceptions import ExportStreamExhausted
+from nucliadb.export_import.models import ExportedItemType, ExportItem
 from nucliadb_protos import knowledgebox_pb2 as kb_pb2
 from nucliadb_protos import resources_pb2, writer_pb2
 from nucliadb_utils.const import Streams
@@ -124,15 +127,19 @@ async def set_labels(
     await ldm.set_labels(kbid, labels)
 
 
-async def iter_broker_messages(
+async def iter_kb_resource_uuids(
     context: ApplicationContext, kbid: str
-) -> AsyncGenerator[writer_pb2.BrokerMessage, None]:
+) -> AsyncGenerator[str, None]:
     rdm = ResourcesDataManager(context.kv_driver, context.blob_storage)
     async for rid in rdm.iterate_resource_ids(kbid):
-        bm = await rdm.get_broker_message(kbid, rid)
-        if bm is None:
-            continue
-        yield bm
+        yield rid
+
+
+async def get_broker_message(
+    context: ApplicationContext, kbid: str, rid: str
+) -> Optional[writer_pb2.BrokerMessage]:
+    rdm = ResourcesDataManager(context.kv_driver, context.blob_storage)
+    return await rdm.get_broker_message(kbid, rid)
 
 
 def get_cloud_files(bm: writer_pb2.BrokerMessage) -> list[resources_pb2.CloudFile]:
@@ -207,3 +214,149 @@ async def get_entities(context: ApplicationContext, kbid: str) -> kb_pb2.Entitie
 async def get_labels(context: ApplicationContext, kbid: str) -> kb_pb2.Labels:
     ldm = LabelsDataManager(context.kv_driver)
     return await ldm.get_labels(kbid)
+
+
+class EndOfStream(Exception):
+    ...
+
+
+class ExportStream:
+    """
+    Models a stream of export bytes that can be read from asynchronously.
+    """
+
+    def __init__(self, export: BytesIO):
+        self.export = export
+        self.read_bytes = 0
+        self.length = len(export.getvalue())
+
+    async def read(self, n_bytes):
+        """
+        Reads n_bytes from the export stream.
+        Raises ExportStreamExhausted if there are no more bytes to read.
+        """
+        if self.read_bytes >= self.length:
+            raise ExportStreamExhausted()
+        chunk = self.export.read(n_bytes)
+        self.read_bytes += len(chunk)
+        return chunk
+
+
+class IteratorExportStream(ExportStream):
+    """
+    Adapts the parent class to be able to read bytes yielded from an async iterator.
+    """
+
+    def __init__(self, iterator: AsyncIterator[bytes]):
+        self.iterator = iterator
+        self.buffer = b""
+
+    def _read_from_buffer(self, n_bytes: int) -> bytes:
+        value = self.buffer[:n_bytes]
+        self.buffer = self.buffer[n_bytes:]
+        return value
+
+    async def read(self, n_bytes: int) -> bytes:
+        while True:
+            try:
+                if self.buffer != b"" and len(self.buffer) >= n_bytes:
+                    return self._read_from_buffer(n_bytes)
+
+                next_chunk = await self.iterator.__anext__()
+                if next_chunk == b"":
+                    raise EndOfStream()
+
+                self.buffer += next_chunk
+                if len(self.buffer) >= n_bytes:
+                    return self._read_from_buffer(n_bytes)
+                else:
+                    # Need to read another chunk
+                    continue
+
+            except (StopAsyncIteration, EndOfStream):
+                if self.buffer != b"":
+                    return self._read_from_buffer(n_bytes)
+                else:
+                    raise ExportStreamExhausted()
+
+
+class ExportStreamReader:
+    """
+    Async generator that reads from an export stream and
+    yields the deserialized export items ready to be imported.
+    """
+
+    def __init__(self, export_stream: ExportStream):
+        self.stream = export_stream
+
+    @property
+    def read_bytes(self) -> int:
+        return self.stream.read_bytes
+
+    async def seek(self, offset: int):
+        await self.stream.read(offset)
+
+    async def read_type(self) -> ExportedItemType:
+        type_bytes = await self.stream.read(3)
+        return ExportedItemType(type_bytes.decode())
+
+    async def read_item(self) -> bytes:
+        size_bytes = await self.stream.read(4)
+        size = int.from_bytes(size_bytes, byteorder="big")
+        data = await self.stream.read(size)
+        return data
+
+    async def read_binary(
+        self,
+    ) -> tuple[resources_pb2.CloudFile, BinaryStreamGenerator]:
+        data = await self.read_item()
+        cf = resources_pb2.CloudFile()
+        cf.ParseFromString(data)
+
+        binary_size_bytes = await self.stream.read(4)
+        binary_size = int.from_bytes(binary_size_bytes, byteorder="big")
+
+        async def file_chunks_generator(chunk_size: int) -> BinaryStream:
+            bytes_read = 0
+            while True:
+                bytes_to_read = min(binary_size - bytes_read, chunk_size)
+                if bytes_to_read == 0:
+                    break
+                chunk = await self.stream.read(bytes_to_read)
+                yield chunk
+                bytes_read += len(chunk)
+
+        return cf, file_chunks_generator
+
+    async def read_bm(self) -> writer_pb2.BrokerMessage:
+        data = await self.read_item()
+        bm = writer_pb2.BrokerMessage()
+        bm.ParseFromString(data)
+        return bm
+
+    async def read_entities(self) -> kb_pb2.EntitiesGroups:
+        data = await self.read_item()
+        entities = kb_pb2.EntitiesGroups()
+        entities.ParseFromString(data)
+        return entities
+
+    async def read_labels(self) -> kb_pb2.Labels:
+        data = await self.read_item()
+        labels = kb_pb2.Labels()
+        labels.ParseFromString(data)
+        return labels
+
+    async def iter_items(self) -> AsyncGenerator[ExportItem, None]:
+        while True:
+            try:
+                item_type = await self.read_type()
+                read_data_func = {
+                    ExportedItemType.RESOURCE: self.read_bm,
+                    ExportedItemType.BINARY: self.read_binary,
+                    ExportedItemType.ENTITIES: self.read_entities,
+                    ExportedItemType.LABELS: self.read_labels,
+                }[item_type]
+                data = await read_data_func()  # type: ignore
+                yield item_type, data
+            except ExportStreamExhausted:
+                break
