+++ nucliadb/nucliadb/writer/api/v1/__init__.py
@@ -17,8 +17,8 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
+from . import export_import  # noqa
 from . import field  # noqa
-from . import import_kb  # noqa
 from . import knowledgebox  # noqa
 from . import resource  # noqa
 from . import services  # noqa
