+++ nucliadb/nucliadb/writer/api/v1/export_import.py
@@ -17,15 +17,31 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
+from uuid import uuid4
+
 from fastapi_versioning import version
 from starlette.requests import Request
 
 from nucliadb.common.context.fastapi import get_app_context
 from nucliadb.export_import import importer
-from nucliadb.export_import.fastapi import FastAPIExportStream
+from nucliadb.export_import.utils import IteratorExportStream
 from nucliadb.writer.api.v1.router import KB_PREFIX, api
+from nucliadb_models.export_import import CreateExportResponse, CreateImportResponse
 from nucliadb_models.resource import NucliaDBRoles
 from nucliadb_utils.authentication import requires_one
+
+
+@api.post(
+    f"/{KB_PREFIX}/{{kbid}}/export",
+    status_code=200,
+    name="Start an export of a Knowledge Box",
+    tags=["Knowledge Boxes"],
+)
+@requires_one([NucliaDBRoles.MANAGER, NucliaDBRoles.WRITER])
+@version(1)
+async def start_kb_export_endpoint(request: Request, kbid: str) -> CreateExportResponse:
+    export_id = uuid4().hex
+    return CreateExportResponse(export_id=export_id)
 
 
 @api.post(
@@ -36,10 +52,19 @@ from nucliadb_utils.authentication import requires_one
 )
 @requires_one([NucliaDBRoles.MANAGER, NucliaDBRoles.WRITER])
 @version(1)
-async def import_kb_endpoint(request: Request, kbid: str) -> None:
+async def import_kb_endpoint(request: Request, kbid: str) -> CreateImportResponse:
     context = get_app_context(request.app)
+    import_id = uuid4().hex
+    stream = FastAPIExportStream(request)
     await importer.import_kb(
         context=context,
         kbid=kbid,
-        stream=FastAPIExportStream(request),
+        stream=stream,
     )
+    return CreateImportResponse(import_id=import_id)
+
+
+class FastAPIExportStream(IteratorExportStream):
+    def __init__(self, request: Request):
+        iterator = request.stream().__aiter__()
+        super().__init__(iterator)
