+++ nucliadb_models/VERSION
@@ -1,1 +1,1 @@
-2.24.0
+2.24.1
