+++ nucliadb_models/nucliadb_models/__init__.py
@@ -22,6 +22,7 @@ import logging
 from .common import *  # noqa
 from .conversation import *  # noqa
 from .datetime import *  # noqa
+from .export_import import *  # noqa
 from .extracted import *  # noqa
 from .file import *  # noqa
 from .keywordset import *  # noqa
