+++ nucliadb_models/nucliadb_models/export_import.py
@@ -17,14 +17,27 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
+from enum import Enum
 
-from nucliadb_models.resource import NucliaDBRoles
-
-DUMMY_EXPORT = b"ENT\x00\x00\x00\x00LAB\x00\x00\x00\x00"
+from pydantic import BaseModel
 
 
-async def test_import_kb(writer_api, knowledgebox_ingest):
-    kbid = knowledgebox_ingest
-    async with writer_api(roles=[NucliaDBRoles.WRITER]) as client:
-        resp = await client.post(f"/kb/{kbid}/import", content=DUMMY_EXPORT)
-        assert resp.status_code == 200
+class CreateExportResponse(BaseModel):
+    export_id: str
+
+
+class CreateImportResponse(BaseModel):
+    import_id: str
+
+
+class Status(str, Enum):
+    SCHEDULED = "scheduled"
+    RUNNING = "running"
+    FINISHED = "finished"
+    FAILED = "failed"
+    ERRORED = "errored"
+    CANCELLED = "cancelled"
+
+
+class StatusResponse(BaseModel):
+    status: Status
